Fix RTSP live streaming and black artifacts
Non-IDR NAL Units were handled as keyframes although they aren't.
This resulted in black screens with partial movement artifacts, e.g.
when streaming started as the non-keyframes were used although the
video information was not complete.
Sometimes the non-keyframes (handled as keyframes) also overwrote the
keyframes so that for tens of seconds only a black screen was visible.